Move `PortfolioFormFields` into `BucketManager`
